Fetch prices from SQS (#2901)
* split price calculation into providers

* add sidecar prices query

* implement coingecko provider

* more cleanup

* cleanup 3

* create data loader on every request

* cache batch loaders

* remove check

* use sidecar for pools

* return 0 if coingecko query fails

* increase batch size

* bump batch size

* send individual queries

* fix build

* fix query

* fix SQS client
